seperate optimize UT and improve UT infra
