+++ .github/workflows/script/unitTest/run_unit_test_optimize.sh
@@ -21,7 +21,7 @@ function pytest() {
     mkdir -p ${coverage_log_dir}
     pip install --no-cache-dir protobuf==3.20.0
 
-    cd /intel-extension-for-transformers/tests || exit 1
+    cd /intel-extension-for-transformers/CI/tests || exit 1
     JOB_NAME=unit_test
     ut_log_name=${LOG_DIR}/${JOB_NAME}.log
     export GLOG_minloglevel=2
@@ -53,7 +53,7 @@ function pytest() {
 }
 
 function main() {
-    bash /intel-extension-for-transformers/.github/workflows/script/unitTest/env_setup.sh
+    bash /intel-extension-for-transformers/.github/workflows/script/unitTest/env_setup.sh "/intel-extension-for-transformers/tests"
     echo "test on ${test_name}"
     if [[ $test_name == "PR-test" ]]; then
         pytest "${LOG_DIR}/coverage_pr"
