+++ intel_extension_for_transformers/neural_chat/tests/requirements.txt
@@ -49,3 +49,4 @@ protobuf==3.20.2
 einops
 urllib3
 langid
+optimum-intel
